+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,20 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Ренат Саляхов on 13.09.2023.
+//
+
+import Foundation
+
+
+struct QuizQuestion {
+  // строка с названием фильма,
+  // совпадает с названием картинки афиши фильма в Assets
+  let image: String
+  // строка с вопросом о рейтинге фильма
+  let text: String
+  // булевое значение (true, false), правильный ответ на вопрос
+  let correctAnswer: Bool
+}
+
