+++ MovieQuiz/Models/QuizResultsViewModel .swift
@@ -1,0 +1,18 @@
+//
+//  QuizResultsViewModel .swift
+//  MovieQuiz
+//
+//  Created by Ренат Саляхов on 13.09.2023.
+//
+
+import Foundation
+
+
+struct QuizResultsViewModel {
+    // строка с заголовком алерта
+    let title: String
+    // строка с текстом о количестве набранных очков
+    let text: String
+    // текст для кнопки алерта
+    let buttonText: String
+}
